+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - inline code spans can now handle complex cases, e.g ``` ``inline  ` code`` ```
     - strike through support
     - Commonmark compliant indented code blocks
+- fixed broken note folder and script moving via drag and drop on macOS
 
 ## 20.1.12
 - fixed a memory leak in the preview (for [#1594](https://github.com/pbek/QOwnNotes/pull/1594), thank you @Waqar144)